+++ src/vs/workbench/contrib/chat/browser/agentSessions/agentSessionsControl.ts
@@ -28,7 +28,7 @@ import { IListStyles } from '../../../../../base/browser/ui/list/listWidget.js';
 import { IStyleOverride } from '../../../../../platform/theme/browser/defaultStyles.js';
 import { IAgentSessionsControl, IMarshalledChatSessionContext } from './agentSessions.js';
 import { HoverPosition } from '../../../../../base/browser/ui/hover/hoverWidget.js';
-import { openSession } from './agentSessionsActions.js';
+import { openSession } from './agentSessionsOpener.js';
 
 export interface IAgentSessionsControlOptions {
 	readonly overrideStyles?: IStyleOverride<IListStyles>;
