+++ src/vs/workbench/contrib/chat/browser/agentSessions/agentSessionsPicker.ts
@@ -10,7 +10,7 @@ import { ThemeIcon } from '../../../../../base/common/themables.js';
 import { localize } from '../../../../../nls.js';
 import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
 import { IQuickInputService, IQuickPickItem, IQuickPickSeparator } from '../../../../../platform/quickinput/common/quickInput.js';
-import { openSession } from './agentSessionsActions.js';
+import { openSession } from './agentSessionsOpener.js';
 import { IAgentSession } from './agentSessionsModel.js';
 import { IAgentSessionsService } from './agentSessionsService.js';
 import { AgentSessionsSorter } from './agentSessionsViewer.js';
